pyproject.toml
[tool.poetry]
name = "cytoid-fastapi"
version = "0.1.0"
description = "Repackaging Cytoid's Graphql interface to return data in the form of Arcaea interface"
authors = ["htxzdfunny <catnetwork@foxmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
python-dotenv = "^1.0.1"
gql = {extras = ["all"], version = "^3.5.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
